Modificaciones Urgentes
Se reviso nuevamente la descripcion del taller, en el cual se pudo evidenciar algunas falencias con lo solitado y lo que esta hecho. Se crearon los metodos "TiempoMultiplicar" y "TiempoObtener", los cuales estan hechos para testear el tiempo de los algortimos "Get" y "Multiply". Estos metodos eran los faltantes para completar con la parte 2 del taller

+++ MatricesPocoPobladas/SparseMatrix.h
@@ -11,7 +11,7 @@ public:
 	int get(int xPos, int yPos);
 	void remove(int xPos, int yPos);
 	void printStoredValues();
-	int density();
+	double density();
 	SparseMatrix* multiply(SparseMatrix* second);
 
 	~SparseMatrix();
